chore: use ull for number of tests

+++ tester.cpp
@@ -13,7 +13,7 @@ std::ofstream error("error.txt");
 
 struct Config {
         void (*const sort)(std::vector<int> &);
-        const size_t number_of_tests;
+        const unsigned long long number_of_tests;
         const size_t length;
         const int min;
         const int max;
@@ -21,8 +21,8 @@ struct Config {
 
     public:
         Config(void (*const sort)(std::vector<int> &),
-               const size_t number_of_tests, const size_t length, const int min,
-               const int max, const bool critical)
+               const unsigned long long number_of_tests, const size_t length,
+               const int min, const int max, const bool critical)
             : sort(sort),
               number_of_tests(number_of_tests),
               length(length),
@@ -30,8 +30,8 @@ struct Config {
               max(max),
               critical(critical) {}
         Config(void (*const sort)(std::vector<int> &),
-               const size_t number_of_tests, const size_t length, const int min,
-               const int max)
+               const unsigned long long number_of_tests, const size_t length,
+               const int min, const int max)
             : Config(sort, number_of_tests, length, min, max, false) {}
 };
